Reorganize `matchables` code
Let's use a directory structure that makes it easier to add XML later.
